server/requirements.txt
argon2-cffi==21.3.0
anyio==3.6.2
autopep8==2.0.1
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.93.0
h11==0.14.0
httptools==0.5.0
idna==3.4
prometheus-client==0.15.0
prometheus-fastapi-instrumentator==5.9.1
pycodestyle==2.10.0
pydantic[email]==1.10.4
python-dotenv==0.21.0
python-engineio==4.3.4
pymongo==4.3.3
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
starlette_exporter==0.15.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
numpy==1.24.1
tritonclient[all]==2.23.0
gevent==22.10.2
scipy==1.10.0
soundfile==0.12.1
pydub==0.25.1
pyjwt==2.6.0
motor==3.1.1
pymongo==4.3.3
celery==5.2.7
celery-flower==1.0.1
redis-om==0.1.2
azure-identity==1.12.0
python-ulid==1.1.0
azure-storage-blob==12.16.0
# asgiref
jsonpickle==3.0.1
webrtcvad==2.0.10
torch==2.0.0
torchaudio==2.0.1
mongodb-migrations==1.2.1
yt-dlp>=2023.3.4
SQLAlchemy==2.0.19
sqlalchemy-timescaledb==0.4.1
psycopg2==2.9.6
FastAPI-SQLAlchemy==0.2.1